+++ crates/compiler/src/mml/tokenizer.rs
@@ -45,6 +45,7 @@ pub enum Token {
     Tie,
     Slur,
     ChangeWholeNoteLength,
+    PlaySample,
     PlayMidiNoteNumber,
     SetDefaultLength,
     Rest,
@@ -248,11 +249,10 @@ impl<'a> Tokenizer<'a> {
             // This list must match `next()`.
             b'0'..=b'9' => false,
             b'a'..=b'g' => false,
-            b'!' | b'@' | b'+' | b'-' | b'[' | b':' | b']' | b'^' | b'&' | b'C' | b'n' | b'l'
-            | b'r' | b'w' | b'o' | b'>' | b'<' | b'v' | b'V' | b'p' | b'Q' | b'~' | b'A' | b'G'
-            | b'E' | b't' | b'T' | b'L' | b'%' | b'.' | b',' | b'|' | b'_' | b'{' | b'}' | b'M' => {
-                false
-            }
+            b'!' | b'@' | b'+' | b'-' | b'[' | b':' | b']' | b'^' | b'&' | b'C' | b's' | b'n'
+            | b'l' | b'r' | b'w' | b'o' | b'>' | b'<' | b'v' | b'V' | b'p' | b'Q' | b'~' | b'A'
+            | b'G' | b'E' | b't' | b'T' | b'L' | b'%' | b'.' | b',' | b'|' | b'_' | b'{' | b'}'
+            | b'M' => false,
             c if c.is_ascii_whitespace() => false,
             _ => true,
         }
@@ -399,6 +399,7 @@ impl<'a> Tokenizer<'a> {
             b'^' => one_ascii_token!(Token::Tie),
             b'&' => one_ascii_token!(Token::Slur),
             b'C' => one_ascii_token!(Token::ChangeWholeNoteLength),
+            b's' => one_ascii_token!(Token::PlaySample),
             b'n' => one_ascii_token!(Token::PlayMidiNoteNumber),
             b'l' => one_ascii_token!(Token::SetDefaultLength),
             b'r' => one_ascii_token!(Token::Rest),